Updated CSS to reduce padding and make various minor improvements for display on small screens.

+++ src/WebUI/ClientApp/src/app/shared/publishingdate.pipe.ts
@@ -15,7 +15,7 @@ export class PublishingDatePipe implements PipeTransform {
     (value.getMonth() == yesterday.getMonth()) &&
     (value.getFullYear() == yesterday.getFullYear());
 
-    const month = value.toLocaleString('default', { month: 'long' });
+    const month = value.toLocaleString('default', { month: 'short' });
     const dayOfMonth = value.getDate();
 
     let dayString: string = "";
